pyproject.toml
[project]
name = "btcmap-admin"
version = "0.1.0"
description = "A web app to maintain data held directly in the BTC Map db."
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.0.3",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.9",
    "requests>=2.32.3",
    "flask-session>=0.8.0",
    "geojson>=3.1.0",
    "geojson-rewind>=1.1.0",
    "shapely>=2.0.6",
    "pyproj>=3.7.0",
    "pillow>=11.1.0",
]
